Fix unit test. E2E test still broken.
