+++ execute/plugin_test.go
@@ -14,6 +14,7 @@ import (
 	"go.uber.org/zap/zapcore"
 	"go.uber.org/zap/zaptest/observer"
 
+	"github.com/smartcontractkit/chainlink-ccip/execute/report"
 	"github.com/smartcontractkit/chainlink-common/pkg/hashutil"
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 	"github.com/smartcontractkit/chainlink-common/pkg/merklemulti"
@@ -23,7 +24,6 @@ import (
 	"github.com/smartcontractkit/libocr/offchainreporting2plus/types"
 	libocrtypes "github.com/smartcontractkit/libocr/ragep2p/types"
 
-	"github.com/smartcontractkit/chainlink-ccip/execute/temp"
 	"github.com/smartcontractkit/chainlink-ccip/internal/libs/slicelib"
 	"github.com/smartcontractkit/chainlink-ccip/internal/mocks"
 	"github.com/smartcontractkit/chainlink-ccip/internal/reader"
@@ -256,7 +256,7 @@ func makeTestCommitReport(
 			uint64(i)))
 	}
 
-	report := plugintypes.ExecutePluginCommitDataWithMessages{
+	commitReport := plugintypes.ExecutePluginCommitDataWithMessages{
 		ExecutePluginCommitData: plugintypes.ExecutePluginCommitData{
 			//MerkleRoot:          root,
 			SourceChain:         cciptypes.ChainSelector(srcChain),
@@ -271,14 +271,14 @@ func makeTestCommitReport(
 	// calculate merkle root
 	root := rootOverride
 	if root.IsEmpty() {
-		tree, err := constructMerkleTree(context.Background(), hasher, report)
+		tree, err := report.ConstructMerkleTree(context.Background(), hasher, commitReport)
 		if err != nil {
 			panic(fmt.Sprintf("unable to construct merkle tree: %s", err))
 		}
-		report.MerkleRoot = tree.Root()
+		commitReport.MerkleRoot = tree.Root()
 	}
 
-	return report
+	return commitReport
 }
 
 // assertMerkleRoot computes the source messages merkle root, then computes a verification with the proof, then compares
@@ -544,207 +544,6 @@ func Test_selectReport(t *testing.T) {
 				lastReport := commitReports[len(commitReports)-1]
 				require.ElementsMatch(t, tt.lastReportExecuted, lastReport.ExecutedMessages)
 			}
-		})
-	}
-}
-
-type badHasher struct{}
-
-func (bh badHasher) Hash(context.Context, cciptypes.Message) (cciptypes.Bytes32, error) {
-	return cciptypes.Bytes32{}, fmt.Errorf("bad hasher")
-}
-
-type badTokenDataReader struct{}
-
-func (btdr badTokenDataReader) ReadTokenData(
-	_ context.Context, _ cciptypes.ChainSelector, _ cciptypes.SeqNum,
-) ([][]byte, error) {
-	return nil, fmt.Errorf("bad token data reader")
-}
-
-type badCodec struct{}
-
-func (bc badCodec) Encode(ctx context.Context, report cciptypes.ExecutePluginReport) ([]byte, error) {
-	return nil, fmt.Errorf("bad codec")
-}
-
-func (bc badCodec) Decode(ctx context.Context, bytes []byte) (cciptypes.ExecutePluginReport, error) {
-	return cciptypes.ExecutePluginReport{}, fmt.Errorf("bad codec")
-}
-
-func Test_buildSingleChainReport_Errors(t *testing.T) {
-	lggr := logger.Test(t)
-
-	type args struct {
-		report          plugintypes.ExecutePluginCommitDataWithMessages
-		maxMessages     int
-		hasher          cciptypes.MessageHasher
-		tokenDataReader temp.TokenDataReader
-		codec           cciptypes.ExecutePluginCodec
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr string
-	}{
-		{
-			name:    "wrong number of messages",
-			wantErr: "unexpected number of messages: expected 1, got 2",
-			args: args{
-				report: plugintypes.ExecutePluginCommitDataWithMessages{
-					ExecutePluginCommitData: plugintypes.ExecutePluginCommitData{
-						SequenceNumberRange: cciptypes.NewSeqNumRange(cciptypes.SeqNum(100), cciptypes.SeqNum(100)),
-					},
-					Messages: []cciptypes.Message{
-						{Header: cciptypes.RampMessageHeader{}},
-						{Header: cciptypes.RampMessageHeader{}},
-					},
-				},
-			},
-		},
-		{
-			name:    "wrong sequence numbers",
-			wantErr: "sequence number 102 outside of report range [100 -> 101]",
-			args: args{
-				report: plugintypes.ExecutePluginCommitDataWithMessages{
-					ExecutePluginCommitData: plugintypes.ExecutePluginCommitData{
-						SequenceNumberRange: cciptypes.NewSeqNumRange(cciptypes.SeqNum(100), cciptypes.SeqNum(101)),
-					},
-					Messages: []cciptypes.Message{
-						{
-							Header: cciptypes.RampMessageHeader{
-								SequenceNumber: cciptypes.SeqNum(100),
-							},
-						},
-						{
-							Header: cciptypes.RampMessageHeader{
-								SequenceNumber: cciptypes.SeqNum(102),
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			name:    "source mismatch",
-			wantErr: "unexpected source chain: expected 1111, got 2222",
-			args: args{
-				report: plugintypes.ExecutePluginCommitDataWithMessages{
-					ExecutePluginCommitData: plugintypes.ExecutePluginCommitData{
-						SourceChain:         1111,
-						SequenceNumberRange: cciptypes.NewSeqNumRange(cciptypes.SeqNum(100), cciptypes.SeqNum(100)),
-					},
-					Messages: []cciptypes.Message{
-						{Header: cciptypes.RampMessageHeader{
-							SourceChainSelector: 2222,
-							SequenceNumber:      cciptypes.SeqNum(100),
-						}},
-					},
-				},
-				hasher: badHasher{},
-			},
-		},
-		{
-			name:    "bad hasher",
-			wantErr: "unable to hash message (1234567, 100): bad hasher",
-			args: args{
-				report: plugintypes.ExecutePluginCommitDataWithMessages{
-					ExecutePluginCommitData: plugintypes.ExecutePluginCommitData{
-						SourceChain:         1234567,
-						SequenceNumberRange: cciptypes.NewSeqNumRange(cciptypes.SeqNum(100), cciptypes.SeqNum(100)),
-					},
-					Messages: []cciptypes.Message{
-						{Header: cciptypes.RampMessageHeader{
-							SourceChainSelector: 1234567,
-							SequenceNumber:      cciptypes.SeqNum(100),
-						}},
-					},
-				},
-				hasher: badHasher{},
-			},
-		},
-		{
-			name:    "bad token data reader",
-			wantErr: "unable to read token data for message 100: bad token data reader",
-			args: args{
-				report: plugintypes.ExecutePluginCommitDataWithMessages{
-					ExecutePluginCommitData: plugintypes.ExecutePluginCommitData{
-						SourceChain:         1234567,
-						SequenceNumberRange: cciptypes.NewSeqNumRange(cciptypes.SeqNum(100), cciptypes.SeqNum(100)),
-					},
-					Messages: []cciptypes.Message{
-						{Header: cciptypes.RampMessageHeader{
-							SourceChainSelector: 1234567,
-							SequenceNumber:      cciptypes.SeqNum(100),
-						}},
-					},
-				},
-				tokenDataReader: badTokenDataReader{},
-			},
-		},
-		{
-			name:    "bad codec",
-			wantErr: "unable to encode report: bad codec",
-			args: args{
-				report: plugintypes.ExecutePluginCommitDataWithMessages{
-					ExecutePluginCommitData: plugintypes.ExecutePluginCommitData{
-						SourceChain:         1234567,
-						SequenceNumberRange: cciptypes.NewSeqNumRange(cciptypes.SeqNum(100), cciptypes.SeqNum(100)),
-					},
-					Messages: []cciptypes.Message{
-						{Header: cciptypes.RampMessageHeader{
-							SourceChainSelector: 1234567,
-							SequenceNumber:      cciptypes.SeqNum(100),
-						}},
-					},
-				},
-				codec: badCodec{},
-			},
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			// Select hasher mock.
-			var resolvedHasher cciptypes.MessageHasher
-			if tt.args.hasher != nil {
-				resolvedHasher = tt.args.hasher
-			} else {
-				resolvedHasher = mocks.NewMessageHasher()
-			}
-
-			// Select token data reader mock.
-			var resolvedTokenDataReader temp.TokenDataReader
-			if tt.args.tokenDataReader != nil {
-				resolvedTokenDataReader = tt.args.tokenDataReader
-			} else {
-				resolvedTokenDataReader = tdr{}
-			}
-
-			// Select codec mock.
-			var resolvedCodec cciptypes.ExecutePluginCodec
-			if tt.args.codec != nil {
-				resolvedCodec = tt.args.codec
-			} else {
-				resolvedCodec = mocks.NewExecutePluginJSONReportCodec()
-			}
-
-			ctx := context.Background()
-			msgs := make(map[int]struct{})
-			for i := 0; i < len(tt.args.report.Messages); i++ {
-				msgs[i] = struct{}{}
-			}
-			execReport, size, _, err := buildSingleChainReport(
-				ctx, lggr, resolvedHasher, resolvedTokenDataReader, resolvedCodec, tt.args.report, msgs)
-			if tt.wantErr != "" {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), tt.wantErr)
-				return
-			}
-			require.NoError(t, err)
-			fmt.Println(execReport, size, err)
 		})
 	}
 }
